Create Loops
1.1 Numbers from 0 to 9
Use a while loop to print out values from 0 to 9.

1.2 Ten Random numbers
Use a while loop to print ten random numbers.

1.3 Sine values from 0 to 360
 

In the last lab, we used trigonometry (SOHCAHTOA) to convert angles and distances to changes in x and y (easting and northing).  We'll the same approach in this assignment (and you can copy the code from the previous assignment).  You can have some fun with this by creating a loop that goes from 0 to 360 printing the x values generated by the equations above.  Then, copy the x values into MS-Excel and create a graph that shows a sine wave.  Remember you'll need to convert the angles from degrees to radians before passing values into the trig functions.

1.4 Sine and Cosine values from 0 to 360
Make a copy of the code above and then add a calculation for the changes in y.  Print the x and y values separated by a comma and then copy and paste these values into Excel.  Plot them in a scattergram or bring them into ArcGIS as "XY" data and you'll see you've created a circle. 

1.5 Your own loop #1
For this part, you can change the value of the "Distance" based on the Angle in the previous part or create a new loop entirely on your own.

1.6 Your own loop #2
Now have some fun and create your own spirograph image using sine and cosine functions.