Create Filtering Lists
1. Finding names
Create a list of at least 10 names.  Create a loop that goes through the names and prints out one specific name from the list.

2. Grouping numbers
Create a loop that generates random numbers between 1 and 100 and counts the number of times a number is over 75, over 50 and less than or equal to 75, over 25 and less than or equal to 50, or less than 25.  Print out the counters.  Increase the number of times through the loop to at least 1000 to see if the numbers make sense.

3. Looking for negative tides
Create a loop that will process a list of values for sea level in Arcata Bay.  The NOAA web site (Links to an external site.) gives these values on a daily basis.  Your task will be to filter an array of these values such as:

SeaLevel=[7.5,2.53,5.86,1.54,7.88,1.85,5.57]

Your loop should print out any values that are at 0 or negative to indicate really low tides.  Your loop should also print out a warning for any sea levels over 8 feet as the dykes may be at risk of breaking at this level.

4. Your own list
Find a list on the Internet and then create a loop to filter the list for things you are interested in.